Move file/couchdb access in separate file / classes
This also uses type annotations in hope for a nicer interface and better error messages
